# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/user/check": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "summary": "检查用户名是否可用",
                "parameters": [
                    {
                        "description": "结构体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CheckForm"
                        }
                    }
                ]
            }
        },
        "/api/user/info": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ]
            }
        },
        "/api/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "结构体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SignForm"
                        }
                    }
                ]
            }
        },
        "/api/user/signup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户相关"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "结构体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SignForm"
                        }
                    }
                ]
            }
        }
    },
    "definitions": {
        "handler.CheckForm": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
        "handler.SignForm": {
            "type": "object",
            "required": [
                "pass",
                "user_id"
            ],
            "properties": {
                "pass": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}